test(core): split dogfood profile responsibilities
